requirements.txt
absl-py==0.13.0
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
idna==2.10
pycodestyle==2.7.0
Pygments==2.9.0
requests==2.25.1
rich==10.4.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
